+++ app/components/sheet-signs.vue
@@ -9,9 +9,9 @@
 				<text v-if="sign.glyph" class="font-char" text-anchor="start" v-html="sign.glyph"></text>
 				<path v-if="sign.def.type === 'path' && !sign.glyph" :d="sign.def.d" :stroke-width="sign.def['stroke-width']" />
 				<rect v-if="sign.def.type === 'rect'"
-					x="0" y="0"
+					x="0" :y="sign.def.height >= 0 ? 0 : sign.def.height"
 					:width="sign.def.width"
-					:height="sign.def.height"
+					:height="Math.abs(sign.def.height)"
 				/>
 				<line v-if="sign.def.type === 'line'"
 					x1="0" y1="0"